requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
asyncmy==0.2.10
attrs==25.3.0
bcrypt==4.3.0
bidict==0.23.1
colorama==0.4.6
colorlog==6.9.0
filetype==1.2.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
pillow==11.1.0
propcache==0.3.1
PyJWT==2.10.1
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
wsproto==1.2.0
yarl==1.19.0
uvloop==0.21.0
aiosmtplib==4.0.0
aiofiles==24.1.0
packaging==24.2